+++ src/pickup.h
@@ -12,26 +12,10 @@ class JsonOut;
 class map;
 struct tripoint;
 
-// TODO: Move this!
-#include "item_location.h"
-#include "optional.h"
-
 namespace pickup
 {
 
-struct pick_drop_selection {
-    item_location target;
-    cata::optional<int> quantity;
-    std::vector<item_location> children;
-
-    void serialize( JsonOut &jsout ) const;
-    void deserialize( JsonIn &jin );
-};
-
-// TODO: This should get information on whether children are consecutive
-/** Finds possible parent-child relations in picked up items to save moves */
-std::vector<pick_drop_selection> optimize_pickup( const std::vector<item_location> &targets,
-        const std::vector<int> &quantities );
+struct pick_drop_selection;
 
 /**
  * Returns `false` if the player was presented a prompt and decided to cancel the pickup.
@@ -62,6 +46,6 @@ int cost_to_move_item( const Character &who, const item &it );
  * @param m map they are on
  */
 bool handle_spillable_contents( Character &c, item &it, map &m );
-} // namespace Pickup
+} // namespace pickup
 
 #endif // CATA_SRC_PICKUP_H
